+++ file.json
@@ -1,1 +1,1 @@
-{"User.0358d372-13fc-41ee-ad89-08a7d6364074": {"id": "0358d372-13fc-41ee-ad89-08a7d6364074", "created_at": "2023-04-30T15:53:53.383956", "updated_at": "2023-04-30T15:53:53.383962", "__class__": "User"}}
+{"Amenity.97ef918a-f9de-4cf3-adcb-4ec452fdd38a": {"id": "97ef918a-f9de-4cf3-adcb-4ec452fdd38a", "created_at": "2023-04-30T16:59:59.710655", "updated_at": "2023-04-30T16:59:59.710655", "__class__": "Amenity"}, "BaseModel.83b83445-130a-4ab3-9aac-09859ee41308": {"id": "83b83445-130a-4ab3-9aac-09859ee41308", "created_at": "2023-04-30T16:59:59.710671", "updated_at": "2023-04-30T16:59:59.710671", "__class__": "BaseModel"}, "City.7e678470-27fc-45e7-869e-02e774f2b11b": {"id": "7e678470-27fc-45e7-869e-02e774f2b11b", "created_at": "2023-04-30T16:59:59.710705", "updated_at": "2023-04-30T16:59:59.710705", "__class__": "City"}, "Place.252d5655-cc60-41e9-a2b0-947d619b5acf": {"id": "252d5655-cc60-41e9-a2b0-947d619b5acf", "created_at": "2023-04-30T16:59:59.710722", "updated_at": "2023-04-30T16:59:59.710722", "__class__": "Place"}, "Review.dc153f6c-9476-4625-b9bb-eb7d3194f35f": {"id": "dc153f6c-9476-4625-b9bb-eb7d3194f35f", "created_at": "2023-04-30T16:59:59.710737", "updated_at": "2023-04-30T16:59:59.710737", "__class__": "Review"}, "State.bbf1c0c2-f61b-4232-8c57-6f976c31f731": {"id": "bbf1c0c2-f61b-4232-8c57-6f976c31f731", "created_at": "2023-04-30T16:59:59.710753", "updated_at": "2023-04-30T16:59:59.710753", "__class__": "State"}, "User.6c4e52ce-cd33-4372-8be0-8e36609727f0": {"id": "6c4e52ce-cd33-4372-8be0-8e36609727f0", "created_at": "2023-04-30T16:59:59.710768", "updated_at": "2023-04-30T16:59:59.710768", "__class__": "User"}}
