+++ tests/test_models/test_engine/test_db_storage.py
@@ -19,6 +19,7 @@ import os
 import pep8
 import pycodestyle
 import unittest
+from models import storage
 DBStorage = db_storage.DBStorage
 classes = {"Amenity": Amenity, "City": City, "Place": Place,
            "Review": Review, "State": State, "User": User}
@@ -71,6 +72,25 @@ test_db_storage.py'])
 
 class TestFileStorage(unittest.TestCase):
     """Test the FileStorage class"""
+    def test_get_return(self):
+        """This method tests the get method of the storage engine"""
+        fake = State(name='fake')
+        storage.new(fake)
+        self.assertIs(type(storage.get(State, fake.id)), type(fake))
+        self.assertTrue(storage.get(State, fake.id) is fake)
+        self.assertTrue(storage.get(State, fake.id) == fake)
+        self.assertIsNone(storage.get(State, 'hello world'))
+
+    def test_count_return(self):
+        """This method tests the count method of the storage engine"""
+        self.assertIs(type(storage.count()), int)
+        self.assertIs(type(storage.count()), type(storage.count(State)))
+        count1 = storage.count(State)
+        fake = State(name='fake')
+        storage.new(fake)
+        count2 = storage.count(State)
+        self.assertTrue(count2 == count1 + 1)
+
     @unittest.skipIf(models.storage_t != 'db', "not testing db storage")
     def test_all_returns_dict(self):
         """Test that all returns a dictionaty"""
